feat: add `orders`, `profile` & move some funcs
