+++ src/pywmapi/profile/__init__.py
@@ -1,0 +1,2 @@
+from .api import *
+from .models import *
