+++ test/test_profile_api.py
@@ -1,0 +1,6 @@
+from pywmapi.common import *
+from pywmapi.profile import *
+
+
+def test_get_profile_by_username():
+    get_profile_by_username("AyajiLin")
